+++ test/serialize/Undefined.js
@@ -1,0 +1,6 @@
+const test = require('ava')
+const { serialize } = require('../..')
+
+test('serialize: Undefined', assert => {
+  assert.deepEqual(serialize({ type: 'Identifier', name: 'undefined' }), undefined)
+})